split pages.py
