+++ zyte_common_items/pages/__init__.py
@@ -1,0 +1,30 @@
+# flake8: noqa
+from .article import ArticlePage, AutoArticlePage, BaseArticlePage
+from .article_list import ArticleListPage, AutoArticleListPage, BaseArticleListPage
+from .article_navigation import (
+    ArticleNavigationPage,
+    AutoArticleNavigationPage,
+    BaseArticleNavigationPage,
+)
+from .base import BasePage, Page
+from .business_place import (
+    AutoBusinessPlacePage,
+    BaseBusinessPlacePage,
+    BusinessPlacePage,
+)
+from .job_posting import AutoJobPostingPage, BaseJobPostingPage, JobPostingPage
+from .mixins import DescriptionMixin, HasMetadata, MetadataT, PriceMixin
+from .product import AutoProductPage, BaseProductPage, ProductPage
+from .product_list import AutoProductListPage, BaseProductListPage, ProductListPage
+from .product_navigation import (
+    AutoProductNavigationPage,
+    BaseProductNavigationPage,
+    ProductNavigationPage,
+)
+from .real_estate import AutoRealEstatePage, BaseRealEstatePage, RealEstatePage
+from .search_request_template import SearchRequestTemplatePage
+from .social_media_post import (
+    AutoSocialMediaPostPage,
+    BaseSocialMediaPostPage,
+    SocialMediaPostPage,
+)
